requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
coloredlogs==15.0.1
colorlog==6.9.0
frozenlist==1.7.0
greenlet==3.2.4
humanfriendly==10.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
SQLAlchemy==2.0.42
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
